refactor: renaming completion
