+++ packages/cli/src/util/runner.ts
@@ -2,7 +2,7 @@ import { IPrismHttpServer } from '@stoplight/prism-http-server/src/types';
 import * as chokidar from 'chokidar';
 import * as os from 'os';
 import { CreateMockServerOptions } from './createServer';
-import { getHttpOperationsFromResource } from '../getHttpOperations';
+import { getHttpOperationsFromResource } from '../operations';
 
 export type CreatePrism = (options: CreateMockServerOptions) => Promise<IPrismHttpServer | void>;
